fix(main,gate): unbreak the build, clear the ratchet, make the bench gate real
`main` was red on all four `ci` steps. Three separate problems, and the third is
the one that let the other two land.

**1. It did not compile.** #957's merge dropped a struct field while keeping
every use of it, and duplicated a match arm:

* `SessionMemory::execution_id` lost its declaration, while `set_execution_id`,
  the constructor, and `learning.rs` all still referenced it (E0609 ×3, E0560).
  Restored with the doc comment it had in d9f9da9c, which explains why it exists
  — without it `execution_reflection.self_rating` is NULL on every row.
* `main.rs` gained a second `Command::Context` arm calling `context_cmd::run`,
  a function that does not exist; the surviving arm calls `run_context` (E0425).
  Removed the dead duplicate and kept one arm, with a comment that describes
  what `run_context` actually does now — the older `#714 lifecycle ledger only`
  wording had been overtaken by the proposal/review commands.
* The same merge duplicated `Command::Context { .. }` in the unreachable-arms
  match, which `-D warnings` rejects.

**2. The file-size ratchet was failing, and baselining it was not allowed.**
`theme.rs` (1602) was over the hard 1500-line limit and explicitly not
grandfathered, so the gate refuses a baseline entry for it. Both it and
`model.rs` (1802, over its ceiling) carried large `#[cfg(test)] mod tests`
blocks, so both are split to `<module>/tests.rs` following the convention
already used by `settings`, `registry`, and `agent`:

    theme.rs  1602 -> 877   (+ theme/tests.rs, 26 tests, all pass)
    model.rs  1802 -> 1130  (+ model/tests.rs, 33 tests, all pass)

Pure relocation — no test changed, added, or removed. Because both now sit under
1500, this **retires** their exemptions rather than raising a ceiling;
`make file-size-update` also drops the now-obsolete `config.rs` and `main.rs`
entries and tightens four others. The baseline moves in the direction the
ratchet exists to enforce.

**3. The bench gate was decoration, and could not simply be promoted.**
#659 moved `harbor==0.6.1` — an audited constant `secure_launcher` refuses to
start without — `bench.yml` ran, reported FAILURE, and the PR merged twelve
minutes later, because the check is not in `main`'s required contexts. It stayed
broken five days.

Promoting it as-is would have deadlocked the repository: the workflow was
filtered with `on.pull_request.paths`, so it does not run when nothing matches,
its check never reports, and a required context that never reports blocks the PR
forever. The filter therefore moves from the trigger into a step. The job now
runs on every PR and always reports; only the two pytest suites are conditional,
so an ordinary Rust PR pays one runner start-up rather than twelve minutes of
Python. That makes `harbor_adapter + analyzer pytest` safe to add to the required
contexts, which is the point.

Verified locally: `check-file-size` OK, `cargo fmt --check` clean, `cargo clippy
--workspace --all-targets -D warnings` clean, `cargo test --workspace` 4679
passed / 0 failed.
