+++ scripts/file-size-baseline.txt
@@ -19,8 +19,6 @@
 2292 stella-cli/src/agent.rs
 1576 stella-cli/src/candidate_ws.rs
 4578 stella-cli/src/command_deck.rs
-1518 stella-cli/src/config.rs
-1548 stella-cli/src/main.rs
 2095 stella-core/src/bus.rs
 2138 stella-core/src/driver.rs
 2820 stella-core/src/driver/tests.rs
@@ -28,18 +26,17 @@
 1609 stella-model/src/anthropic/tests.rs
 1780 stella-model/src/bedrock.rs
 1948 stella-model/src/openai.rs
-1738 stella-model/src/zai/tests.rs
+1684 stella-model/src/zai/tests.rs
 2626 stella-pipeline/src/pipeline.rs
 2159 stella-pipeline/src/pipeline/tests.rs
 2566 stella-protocol/src/event.rs
-2292 stella-store/src/lib.rs
-2269 stella-store/src/tests.rs
+2055 stella-store/src/lib.rs
+2204 stella-store/src/tests.rs
 1892 stella-store/src/usage.rs
 1566 stella-tools/src/media.rs
 3271 stella-tools/src/registry.rs
 1837 stella-tools/src/scripts.rs
-1782 stella-tui/src/deck_render.rs
-4009 stella-tui/src/deck_ui.rs
-1761 stella-tui/src/model.rs
+1723 stella-tui/src/deck_render.rs
+3999 stella-tui/src/deck_ui.rs
 1660 stella-tui/src/views/engine.rs
-1525 stella-tui/src/views/session.rs
+1524 stella-tui/src/views/session.rs
